Add cache comparison
Not entirely sure if it makes all the sense the way it is. Might want
to compare all paths for both.

However, the main inspiration has come from an idea to cache
calculated data, like a csv. If that csv is recalculated
at some point and something depends on that csv, the comparison
can be made based on the current state and what's in the cache
on file. If something is not cached on file but is set as a dependency,
that means that the
dependent calculation needs to be run regardless. In turn, if
something is cached on file but nothing
depends on that data, then it is essentially irrelevant for the moment.
As such, the current functionality makes sense for the intended purpose.