Change to using shelve module for caching with FunctionCache
Commit messages from oldest to newest:
--------------------------------------

Make abstract cache more abstract, create separate JSON cache

Originally went with the JSON angle for the abstract cache (making
it not very abstract anyway),
but as I'm thinking of changing to using the shelve module instead,
I felt it best to keep the JSON stuff but move it to a separate class,
and make a child class likewise for the shelve stuff.

Add utility functions for shelve saving and loading

Make cache loading OR database creation automatic

Add cacher that uses shelve module

Make function_cache more clearly structured

Kind of flip-flopping on this NamedTuple or TypedDict thing.
Though actually, the only problem is if you want to directly
get the container as a set of key-value pairs, as I don't think
the NamedTuple does that (put it into a JSON parser and the parser considers
it a tuple, discarding the keys, effectively). Can obviously do ._asdict(),
but that requires some custom logic to be possibly added, say, if you
just want to quickly JSON some highly nested container with a NamedTuple in
it. The tuple works better if you don't have to serialise, though.

Add tests for function cache and inspect utilities

Sorta concerned as for some reason the function cache tests
failed at first ("OSError: could not get source code"),
yet after no meaningful change in them, they seemed to work and have
since been working. If there is a problem, it would be with the getsource
inspect function not somehow being able to get the function, at least
based on the error.