Perform refactoring
Felt like maybe the names are a little confusing. Decided that
"cacher" makes more sense: the outer object contains a cache, so it's
difficult to refer to both the inner and outer object as "cache".
As the outer object does the actual work, it also makes more sense
to call it the thing that does the caching (the "cacher"), while the cache itself
just contains the cached data.