Add deque cache
The FunctionCacher's cache is a little annoying to work with as is,
and it felt prudent to create something that would be a little
better. This should be fairly general as well, allowing more easily
to find objects in deques, also automatically moves the accessed
object to the front of the cache.