Make DequeCache picklable
Previously, inheriting from defaultdict caused the issue that
it was not possible to pass the factory function to the defaultdict
initialiser and later shelve the DequeCache: it's not possible to
do lambdas (can't be pickled), and passing
in a instance method resulted in an infinite recursion. Changing to
implementing the necessary feature directly avoids the use of a lambda
and the recursion does not become a problem either.