Update to Python 3.13
Mainly wanted to do this as these the ability to use defaults for
generic types, apparently. In turn, I do also want to keep the version
updated as best I can.

Had to change the shelve deletion a little because now my python
installation was suddenly able to work with sqlite3 databases,
whereas previously it was supposedly dbm.dumb. I don't know what
the file extensions for the other two are (dbm.gnu and dbm.ndbm),
though looks like the documentation for the open methods for each
might give some idea.