Make imports relative in main code
When the package is built and imported, the direct references to to the source
folder won't work as the package doesn't exist in a src folder. So,
need to have the imports be relative, ideally. All else should work
fine with the src references as only the src contents should be a part
of the package. Tests, for example, shouldn't exist in the installed
package, so it should be fine to refer to the src directly in them.