Switch to using uv for package management
Came across uv through pdm itself, which has an experimental uv
integration. uv promises much better speed, and thus far seems to deliver.
Besides that, it seems to work basically the same with regard to every
other feature compared to pdm, so seems like decent thing to switch to.

Changed the pyproject file to allow builds too (did this a while ago,
actually, just didn't want to commit yet for whatever reason). Decided
to stick with pdm for the backend as I had already been testing that.
Setuptools would be a sensible option, I guess, as it's related to PyPA,
but it adds some egg-info nonsense to the source folder? Absolute insanity.
Eggs are deprecated https://packaging.python.org/en/latest/discussions/package-formats/#what-about-eggs,
so even less sensical. Anyway, to avoid having to fiddle with all that,
Might as well continue to use pdm as the backend.