Change tests to use package
Rather than importing from source, the tests now import an installed
editable package of the project, or such is the idea. Decided to
change to doing this as uv actually automatically installs the
editable package by the looks of it. Well, the actual reason is of course
that technically testing the installed package is more accurate
as that's largely how it's going to be used in practice. uv just made
it easier to actually do this, where pdm seemed a little weird with
editable installs.