Fix automated package versioning
Not entirely sure on how to actually do this. I want to have the
tests run as a separate workflow to the publish tasks (to avoid repetition,
maybe being unnecessarily conservative with it though?), and in turn
I want the publish tasks to be manual for now and obviously something
that would be run only after the tests have succesfully run. However,
I also want to manage releases (tagging) from Github (maybe this
insistence is the issue), and the publish tasks to automatically pick
their release version as the version that of given by the tags.

The issue is that naturally I can't tag the release before the tests
are run (from the Github side, unless I decide to just use dispatching
to run the tests? Realistically, the tests should pass before a new
release is created, though, so the tests should not assume a current
tag), so they can't actually get the version number properly
the way it was before, so now it needs to just be set as the most
recent tag. Fine as such, seeing as once the tests have run,
it should be possible to create the release and then the publish
tasks would work fine. When forgetting to add a new release, the
new version will use the newest tag, which would likely be in
the PyPI index already so would be rejected, which is ideal.

Actually, in any case I want to be able to run the tests separately
from the publishing, because it might be that I start using pull requests
instead, in which case the tests should obviously pass there. For this,
assuming tags doesn't make really make sense, so having a fallback
version seems sensible.